services/face-swap/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
onnxruntime>=1.19.0
onnx>=1.19.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
pillow>=10.4.0
python-multipart>=0.0.12
